Module vrme_server::accounts::register

Handles user registration and account creation.

We require that the client supply a hashed password from the client-side which shall be generated by a cryptographically-strong hashing function such as SHA-256 to avoid transferring the plain-text password over the network which could be compromised.

Reference

SHA-256

Structs

RegistrationRequest
Required payload when a user wishes to register to create a new account.

Constants

CREATE_USER_QUERY

_IMPL_DESERIALIZE_FOR_RegistrationRequest

_IMPL_SERIALIZE_FOR_RegistrationRequest

Functions

create_account_if_not_exists

handle_registration
POST endpoint handler for user registration.

make_success_response

validate_email

validate_name_length

validate_request_payload