Function vrme_server::accounts::update_info::handle_update_user_account

pub async fn handle_update_user_account(
    pool: Data<PersistentConnectionPool>, 
    path: Path<Uuid>, 
    req: Json<UpdateAccountInformationRequest>
) -> Result<HttpResponse, Error>

Handles endpoint for updating user information at PUT /account/{uuid}.

Path Information

The route parameter required is:

{uuid}: the unique identifier of the user generated upon successful registration.

Authentication Required

Required to supply the Authentication: Bearer <uuid-token-pair>.

Required Payload

Specify Content-Type: application/json and Content-Length.

struct UpdatedAccountInformation {
    first_name: String,
    last_name: String,
}

Example

PUT /account/{uuid}
Authorization: Bearer eyJ1dWlkIjoiMTIzZTQ1NjctZTg5Yi0xMmQzLWE0NTYtNDI2NjU1NDQwMDAwIiwiYXV0aC10b2tlbiI6ImJhc2U2NC1lbmNvZGVkIn0=
Content-Type: application/json

{
    "first_name": "NewFirstName",
    "last_name": "NewLastName"
}

Note

If the uuid of a user is lost by the client, it is possible to recover the uuid of the user by the GET /user/uuid endpoint.