Module vrme_server::auth::login

Handles user login and auth_token issuing.

Structs

LoginInfo
Required login payload – the user needs to login with their email and hashed_password.

Constants

GET_PREVIOUS_HASH_QUERY

UPSERT_AUTH_SESSION_QUERY

_IMPL_DESERIALIZE_FOR_LoginInfo

_IMPL_SERIALIZE_FOR_LoginInfo

Functions

check_registration

handle_login
User login handler.

make_success_response

upsert_auth_session
Either creates a new auth_session, or refreshes an existing session with a new auth_token.