Function vrme_server::auth::middleware::identity_validator

pub async fn identity_validator(
    req: ServiceRequest, 
    credentials: BearerAuth
) -> Result<ServiceRequest, ActixError>

Validator function that filters requests and based on client-provided authentication information (or the lack thereof), decides whether to reject further action (i.e. 401 Unauthorized) or to pass on the handling to further handlers down the response chain.

Note

If the auth_token associated with the uuid exists, but is outdated, the request will be rejected as the client shall POST /login again.

If the auth_token associated with the uuid exists and has not expired yet, the request will be accepted AND the auth_token's validity duration will be refreshed.