Function vrme_server::avatars::upload::handle_upload_avatar

pub async fn handle_upload_avatar(
    auth: BearerAuth, 
    uuid: Path<Uuid>, 
    __arg2: Multipart
) -> Result<HttpResponse, Error>

Handler for avatar image upload.

The endpoint is protected.

The client should upload the image in PNG format – this should probably be done using a HTML form which has a file upload entry, so the client can send the POST payload as a multipart/form-data where the server assumes the first form data is the desired avatar.

Caveats

We do not perform any validation on whether the supplied binary data is in fact valid PNG – it is the client's responsibility to make sure that a valid PNG is sent as the avatar.

If an invalid avatar is uploaded, then anyone who calls GET /account/{uuid}/avatar will get back the invalid avatar.

References

Multipart form specified by RFC 7578