Module vrme_server::meetings::init_session

Handles meeting session initiation.

Structs

MeetingSessionResponsePayload
Response payload upon successful meeting session initialization.

Constants

UPSERT_MEETING_SESSION_QUERY

_IMPL_DESERIALIZE_FOR_MeetingSessionResponsePayload

_IMPL_SERIALIZE_FOR_MeetingSessionResponsePayload

Functions

create_new_session_or_return_existing

handle_init_session
Handler for initializing a meeting session. It must be intiated by an authenticated presenter. If a meeting session associated with the presenter already exists, then info on th existing meeting session is returned; otherwise a fresh meeting session is created and its info returned.