Module vrme_server::meetings::meeting_session

Data type definition for a meeting session.

Structs

MeetingSession
A MeetingSession represents an on-going meeting. A meeting session consists of one presenter and zero or more listeners. Each meeting session is given a unique Uuid to differentiate between sessions and to determine which user(s) have priviledges to change meeting session settings, invite attendees, kick attendees, upload presentation slides, etc.

Type Definitions

Listener
A listener, identified by their user_id.

MeetingId
A meeting's id to differentiate between different meetings.

Presenter
A presenter, identified by their user_id.