Module vrme_server::settings

Settings for the application server.

Supports parsing tiered settings format and overriding from environment variables prefixed with APP_ for settings keys of the same name.

Currently support format:

TOML

YAML

JSON

Structs

AuthSettings
Authentication settings.

DatabaseSettings
Database settings.

LoggingSettings
Logging settings.

RateLimitingSettings
Rate limiting settings.

ServerSettings
Server settings.

Settings
The settings for the application.

TlsSettings
TLS settings.

Enums

LoggingLevel
Logging levels. The most specific is LoggingLevel::Trace, and the least specific is LoggingLevel::Error.

SettingsError
Errors encountered when trying to determine the settings for the application.

Constants

NESTED_SETTINGS_ENV_NAME_SEPARATOR

_IMPL_DESERIALIZE_FOR_AuthSettings

_IMPL_DESERIALIZE_FOR_DatabaseSettings

_IMPL_DESERIALIZE_FOR_LoggingLevel

_IMPL_DESERIALIZE_FOR_LoggingSettings

_IMPL_DESERIALIZE_FOR_RateLimitingSettings

_IMPL_DESERIALIZE_FOR_ServerSettings

_IMPL_DESERIALIZE_FOR_Settings

_IMPL_DESERIALIZE_FOR_TlsSettings

Functions

default_auth_token_length
Default auth_token length in bytes. The auth_token should be strong enough (partly by being long enough, and also from being generated by cryptographically-strong CSPRNG that at periodically samples from true random entropy sources.

default_auth_token_validity_duration
Default auth_token validity duration is 30 days.

default_cooldown_duration
Default cooldown duration is 60 seconds.

default_json_size_limit

default_max_requests
Default max requests per duration is clamped to 100.

default_pool_size