Module vrme_server::types::hashed_password

Hashed password newtype.

Structs

HashedPassword
Hashed password type.

Constants

HASHED_PASSWORD_LEN
Length of the extracted hashed password in bytes. This is for the raw hashed password bytes that is not Base64-encoded.

PBKDF2_ITERATIONS
Number of PBKDF2 iterations to perform. The more iterations, the more difficult to try to compute a rainbow table to try to reverse the hash. However, more iterations also take more CPU cycles to compute.

SALT_LEN
Length of the randomly generated salt in bytes.

_IMPL_DESERIALIZE_FOR_HashedPassword

_IMPL_SERIALIZE_FOR_HashedPassword

Statics

PBKDF2_ALGORITHM
We use the PBKDF2 algorithm to compute the password hash in a secure fashion, with the core hash function being HMAC-SHA-256.